pyproject.toml
[project]
name = "checker-fxxker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "2captcha-python>=1.5.1",
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "mashumaro>=3.16",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "schedule>=1.2.2",
]
